imager/requirements.txt
Django==1.7.5
Pillow==2.7.0
django-debug-toolbar==1.3.0
django-registration-redux==1.1
factory-boy==2.4.1
psycopg2==2.6
sorl-thumbnail==12.2
sqlparse==0.1.14
wsgiref==0.1.2
